feat: new routing engine class for express route registration
The route engine provides the core functionality for registering routes on an express app. It uses a context clue to indicate how the files should be imported into the application

BREAKING CHANGE: This feature introduces a new way to register routes in Express and requires change to existing code by instantiating the new `RouteEngine` class.